+++ stocal/transitions.py
@@ -202,7 +202,7 @@ class Reaction(Transition):
         """
         return 0.
 
-    def next_occurrence(self, time, state):
+    def next_occurrence(self, time, state, rng=None):
         """Determine next reaction firing time.
 
         This is a helper function to use Reactions in next-firing-time
@@ -210,13 +210,14 @@ class Reaction(Transition):
         from a Poisson distribution with mean propensity  and returns
         the given current time plus the delay.
         """
-        from random import random
         from math import log
+        if not rng:
+            import random as rng
         propensity = self.propensity(state)
         if not propensity:
             return float('inf')
         else:
-            return time - log(random())/propensity
+            return time - log(rng.random())/propensity
 
     def propensity_integral(self, state, time, delta_t):
         """Integrate propensity function from time to time+delta_t
@@ -362,7 +363,7 @@ class Event(Transition):
             ))
         return self._hash
 
-    def next_occurrence(self, time, state=None):
+    def next_occurrence(self, time, state=None, rng=None):
         """Next occurrence of the Event at or after time.
 
         If the event does not re-occur, returns float('inf').
@@ -560,7 +561,6 @@ class ReactionRule(Rule):
         for reactants in combinations([], self.signature, novel_species, False):
             for trans in self.novel_reactions(*reactants):
                 yield trans
-
 
 
 class Process(object):