+++ CHANGELOG.md
@@ -1,12 +1,23 @@
 # Changelog
 
-## [1.1.1]
+## [1.2]
+
+### Added
+- Modular trajectory sampling interface stocal.experimental.samplers
+- TrajectorySampler instances accept an optional random seed
+
+### Changed
+- TrajectorySampler instances now use an internal random number generator (sampler.rng) rather than python's global one.
+
+
+## [1.1.1] - 2018-03-21
 
 ### Fixed
 - Fixed issue 3 in AndersonNRM
 
 ### Changed
 - MassAction.__repr__ now also prints rate constant
+
 
 ## [1.1] - 2018-03-08